# 文档/openapi.yaml
openapi: 3.0.3
info:
  title: 体育教培机构管理系统 API
  version: v1.0.0
  description: |
    基于《需求分析文档》《系统架构设计文档》《数据库详细设计文档》《API接口设计文档》生成的 OpenAPI 规范文件。
    本规范覆盖小程序端、后台管理端与第三方回调接口，适用于 Swagger UI、Knife4j、OpenAPI Generator 等工具。
servers:
  - url: http://localhost:8080
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境示例地址
...
security:
  - bearerAuth: []
paths:
  /api/app/v1/auth/wx-login:
    post:
      tags: [AppAuth]
      summary: 微信登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WxLoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/AuthLoginResponse'
...
  /api/app/v1/auth/profile:
    put:
      tags: [AppAuth]
      summary: 更新小程序用户资料
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppProfileUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
...
  /api/app/v1/courses:
    get:
      tags: [AppCourse]
      summary: 课程列表
      security: []
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
        - name: courseType
          in: query
          schema:
            type: integer
        - name: sportType
          in: query
          schema:
            type: string
        - name: priceMin
          in: query
          schema:
            type: number
            format: double
        - name: priceMax
          in: query
          schema:
            type: number
            format: double
        - name: location
          in: query
          schema:
            type: string
        - name: sortType
          in: query
          schema:
            type: string
            enum: [recommend, sales, priceAsc, latest]
      responses:
        '200':
          $ref: '#/components/responses/CoursePageResponse'
  /api/app/v1/courses/{courseId}:
    get:
      tags: [AppCourse]
      summary: 课程详情
      security: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          $ref: '#/components/responses/CourseDetailResponse'
  /api/app/v1/courses/{courseId}/schedules:
    get:
      tags: [AppCourse]
      summary: 团课排期列表
      security: []
      parameters:
        - $ref: '#/components/parameters/CourseId'
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          $ref: '#/components/responses/ScheduleListResponse'
...
  /api/admin/v1/courses:
    get:
      tags: [AdminCourse]
      summary: 后台课程列表
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/CoursePageResponse'
    post:
      tags: [AdminCourse]
      summary: 新增课程
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpsertRequest'
      responses:
        '200':
          $ref: '#/components/responses/CourseDetailResponse'
  /api/admin/v1/courses/{courseId}:
    get:
      tags: [AdminCourse]
      summary: 后台课程详情
      parameters:
        - $ref: '#/components/parameters/CourseId'
      responses:
        '200':
          $ref: '#/components/responses/CourseDetailResponse'
    put:
      tags: [AdminCourse]
      summary: 编辑课程
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpsertRequest'
      responses:
        '200':
          $ref: '#/components/responses/CourseDetailResponse'
  /api/admin/v1/courses/{courseId}/status:
    post:
      tags: [AdminCourse]
      summary: 上下架课程
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusUpdateRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
...
  /api/admin/v1/coaches:
    get:
      tags: [AdminCoach]
      summary: 教练列表
      responses:
        '200':
          $ref: '#/components/responses/CoachPageResponse'
    post:
      tags: [AdminCoach]
      summary: 新增教练
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachUpsertRequest'
      responses:
        '200':
          $ref: '#/components/responses/CoachDetailResponse'
  /api/admin/v1/coaches/{coachId}:
    put:
      tags: [AdminCoach]
      summary: 编辑教练
      parameters:
        - $ref: '#/components/parameters/CoachId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CoachUpsertRequest'
      responses:
        '200':
          $ref: '#/components/responses/CoachDetailResponse'
  /api/admin/v1/coaches/{coachId}/workload:
    get:
      tags: [AdminCoach]
      summary: 教练课时统计
      parameters:
        - $ref: '#/components/parameters/CoachId'
      responses:
        '200':
          $ref: '#/components/responses/CoachWorkloadResponse'
  /api/admin/v1/schedules:
    get:
      tags: [AdminSchedule]
      summary: 团课排期列表
      parameters:
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          $ref: '#/components/responses/SchedulePageResponse'
    post:
      tags: [AdminSchedule]
      summary: 新增排期
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScheduleRequest'
      responses:
        '200':
          $ref: '#/components/responses/ScheduleDetailResponse'
  /api/admin/v1/schedules/{scheduleId}/cancel:
    post:
      tags: [AdminSchedule]
      summary: 取消排期
      parameters:
        - $ref: '#/components/parameters/ScheduleId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelScheduleRequest'
      responses:
        '200':
          $ref: '#/components/responses/SuccessResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    CoachId:
      name: coachId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ScheduleId:
      name: scheduleId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    SuccessResponse:
      description: 成功
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    AuthLoginResponse:
      description: 登录成功
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/LoginResult'
...
    CoursePageResponse:
      description: 课程分页
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CoursePageEnvelope'
    CourseDetailResponse:
      description: 课程详情
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CourseDetail'
    ScheduleListResponse:
      description: 排期列表
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ScheduleListEnvelope'
    SchedulePageResponse:
      description: 排期分页
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SchedulePageEnvelope'
    ScheduleDetailResponse:
      description: 排期详情
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Schedule'
...
    CoachPageResponse:
      description: 教练分页
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CoachPageEnvelope'
    CoachDetailResponse:
      description: 教练详情
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CoachDetail'
    CoachWorkloadResponse:
      description: 教练工作量
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data:
                    $ref: '#/components/schemas/CoachWorkload'
...
  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          nullable: true
          additionalProperties: true
      required: [code, message]
    PageMeta:
      type: object
      properties:
        pageNo:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
      required: [pageNo, pageSize, total]
    LoginResult:
      type: object
      properties:
        token:
          type: string
        userId:
          type: integer
          format: int64
        openid:
          type: string
        nickname:
          type: string
        isNewUser:
          type: boolean
...
    CourseSimple:
      type: object
      properties:
        courseId:
          type: integer
          format: int64
        courseName:
          type: string
        courseType:
          type: integer
        sportType:
          type: string
        coverUrl:
          type: string
        price:
          type: number
          format: double
        originalPrice:
          type: number
          format: double
          nullable: true
        lessonCount:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            type: string
    ProductSimple:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        productName:
          type: string
        coverUrl:
          type: string
        price:
          type: number
          format: double
        originalPrice:
          type: number
          format: double
          nullable: true
    GroupOrderSimple:
      type: object
      properties:
        groupOrderId:
          type: integer
          format: int64
        groupNo:
          type: string
        groupType:
          type: integer
        targetCount:
          type: integer
        currentCount:
          type: integer
        expireTime:
          type: string
          format: date-time
        status:
          type: integer
...
    Schedule:
      type: object
      properties:
        scheduleId:
          type: integer
          format: int64
        courseId:
          type: integer
          format: int64
        coachId:
          type: integer
          format: int64
        scheduleDate:
          type: string
          format: date
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
        minGroupCount:
          type: integer
        enrolledCount:
          type: integer
        status:
          type: integer
    CourseDetail:
      allOf:
        - $ref: '#/components/schemas/CourseSimple'
        - type: object
          properties:
            description:
              type: string
            validityDays:
              type: integer
            isDoorToDoor:
              type: boolean
            serviceAreas:
              type: array
              items:
                type: string
            fixedScheduleDesc:
              type: string
              nullable: true
            fixedLocation:
              type: string
              nullable: true
            maxParticipants:
              type: integer
              nullable: true
            relatedProducts:
              type: array
              items:
                $ref: '#/components/schemas/ProductSimple'
            groupOptions:
              type: array
              items:
                $ref: '#/components/schemas/GroupOrderSimple'
...
    CoachDetail:
      type: object
      properties:
        coachId:
          type: integer
          format: int64
        coachName:
          type: string
        phone:
          type: string
        sportItems:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
          format: double
        availableTimes:
          type: array
          items:
            type: object
            additionalProperties: true
    CoachWorkload:
      type: object
      properties:
        coachId:
          type: integer
          format: int64
        monthLessonCount:
          type: integer
        hourlyRate:
          type: number
          format: double
        payableAmount:
          type: number
          format: double
...
    WxLoginRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
        nickname:
          type: string
        avatarUrl:
          type: string
    AppProfileUpdateRequest:
      type: object
      properties:
        realName:
          type: string
        phone:
          type: string
        parentName:
          type: string
        birthday:
          type: string
          format: date
        gender:
          type: integer
...
    CourseUpsertRequest:
      type: object
      required: [courseName, courseType, sportType, price]
      properties:
        courseName:
          type: string
        courseType:
          type: integer
        sportType:
          type: string
        coverUrl:
          type: string
        description:
          type: string
        detailImages:
          type: array
          items:
            type: string
        price:
          type: number
          format: double
        originalPrice:
          type: number
          format: double
        lessonCount:
          type: integer
        validityDays:
          type: integer
        isDoorToDoor:
          type: boolean
        serviceAreas:
          type: array
          items:
            type: string
        fixedScheduleDesc:
          type: string
        fixedLocation:
          type: string
        maxParticipants:
          type: integer
        groupSuccessCount:
          type: integer
        operationWeight:
          type: integer
    StatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: integer
...
    CoachUpsertRequest:
      type: object
      required: [coachName]
      properties:
        coachName:
          type: string
        phone:
          type: string
        sportItems:
          type: array
          items:
            type: string
        introduction:
          type: string
        hourlyRate:
          type: number
          format: double
        availableTimes:
          type: array
          items:
            type: object
            additionalProperties: true
    CreateScheduleRequest:
      type: object
      required: [courseId, coachId, scheduleDate, startTime, endTime, location, capacity, minGroupCount]
      properties:
        courseId:
          type: integer
          format: int64
        coachId:
          type: integer
          format: int64
        scheduleDate:
          type: string
          format: date
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        location:
          type: string
        capacity:
          type: integer
        minGroupCount:
          type: integer
    CancelScheduleRequest:
      type: object
      properties:
        reason:
          type: string
...
    CoursePageEnvelope:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/CourseSimple'
    ScheduleListEnvelope:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
    SchedulePageEnvelope:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/Schedule'
...
    CoachPageEnvelope:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        data:
          allOf:
            - $ref: '#/components/schemas/PageMeta'
            - type: object
              properties:
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/CoachDetail'